Update the Flexbox Design and Fix Some Issues
In this commit, the white space issue at the end of the footer was
fixed. This issue was as a result of the web page containing little
content not enough to fill up the web page. The issue was fixed using
the Flexbox property, "flex-grow" where the property was set at a value
of 2.