Fix Some Bugs and Improve User Experience
A user identified a bug in the mathematical computation of the total
price of goods bought which extended to also affect the total price of
all goods bought.

This commit fixed that bug hence any changes made by the user during the
user's course of shopping such as removing and adding items to the cart
will reflect appropriately on the neccessary mathematical computation
that follows.

This commit also added some changes that will enhance user experience
and the aesthetics to the website's layout design.